Added navigation between pages
