+++ app/src/main/java/com/example/eventlottery/ui/profilepage/ProfilePage.java
@@ -16,26 +16,42 @@ import android.widget.TextView;
 import com.example.eventlottery.R;
 import com.example.eventlottery.databinding.FragmentProfilePageBinding;
 
+import com.google.firebase.firestore.CollectionReference;
+import com.google.firebase.firestore.FirebaseFirestore;
+
+import android.widget.Button;
+
 public class ProfilePage extends Fragment {
 
     private ProfilePageViewModel mViewModel;
 
     private FragmentProfilePageBinding binding;
 
+    private FirebaseFirestore db;
+
+    private Button saveButton;
+
     public static ProfilePage newInstance() {
         return new ProfilePage();
     }
+
 
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
                              @Nullable Bundle savedInstanceState) {
 
+        db = FirebaseFirestore.getInstance();
+
+        CollectionReference userRef = db.collection("users");
 
         binding = FragmentProfilePageBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
         final TextView textView = binding.fNameTextView;
-        mViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+
+        mViewModel = new ViewModelProvider(this).get(ProfilePageViewModel.class);
+        // TODO: Use the ViewModel
+
         return inflater.inflate(R.layout.fragment_profile_page, container, false);
     }
 
@@ -57,3 +73,20 @@ public class ProfilePage extends Fragment {
         binding = null;
     }
 }
+
+/*
+// This was taken from https://firebase.google.com/docs/firestore/manage-data/add-data#data_types October 18 2024
+        HashMap<String, Object> user_data = new HashMap<>();
+        user_data.put("android_id", androidID);
+        user_data.put("f_name", "Chirayu");
+        user_data.put("l_name", "Shah");
+        user_data.put("email", "cshah1@ualberta.ca");
+        user_data.put("isAdmin", true);
+
+        userRef.document("admin_1").set(user_data).addOnSuccessListener(new OnSuccessListener<Void>() {
+            @Override
+            public void onSuccess(Void unused) {
+                Log.d("Firestore", "Document successfully written");
+            }
+        });
+ */
