+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -22,4 +22,9 @@
         android:name="com.example.eventlottery.ui.notifications.NotificationsFragment"
         android:label="@string/title_notifications"
         tools:layout="@layout/fragment_notifications" />
+    <fragment
+        android:id="@+id/navigation_profile"
+        android:name="com.example.eventlottery.ui.profilepage.ProfilePage"
+        android:label="@string/title_profile"
+        tools:layout="@layout/fragment_profile_page" />
 </navigation>
